+++ src/main/java/ch/danielt/akw/block/MultiblockReactorControllerBlock.java
@@ -22,6 +22,7 @@ import net.minecraft.util.hit.BlockHitResult;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
 import net.minecraft.world.World;
+import net.minecraft.world.block.WireOrientation;
 import org.jetbrains.annotations.Nullable;
 
 /**
@@ -34,18 +35,42 @@ public class MultiblockReactorControllerBlock extends Block implements BlockEnti
     public static final EnumProperty<Direction> FACING = Properties.HORIZONTAL_FACING;
     public static final BooleanProperty ASSEMBLED = BooleanProperty.of("assembled");
     public static final BooleanProperty LIT = Properties.LIT;
+    public static final BooleanProperty POWERED = Properties.POWERED;
 
     public MultiblockReactorControllerBlock(Settings settings) {
         super(settings);
         setDefaultState(getDefaultState()
                 .with(FACING, Direction.NORTH)
                 .with(ASSEMBLED, false)
-                .with(LIT, false));
+                .with(LIT, false)
+                .with(POWERED, false));
     }
 
     @Override
     protected void appendProperties(StateManager.Builder<Block, BlockState> builder) {
-        builder.add(FACING, ASSEMBLED, LIT);
+        builder.add(FACING, ASSEMBLED, LIT, POWERED);
+    }
+
+    @Override
+    protected boolean hasComparatorOutput(BlockState state) {
+        return true;
+    }
+
+    @Override
+    protected int getComparatorOutput(BlockState state, World world, BlockPos pos, Direction direction) {
+        if (world.getBlockEntity(pos) instanceof MultiblockReactorControllerBlockEntity be) {
+            return be.getComparatorLevel();
+        }
+        return 0;
+    }
+
+    @Override
+    protected void neighborUpdate(BlockState state, World world, BlockPos pos, Block sourceBlock,
+                                   WireOrientation wireOrientation, boolean notify) {
+        boolean powered = world.isReceivingRedstonePower(pos);
+        if (powered != state.get(POWERED)) {
+            world.setBlockState(pos, state.with(POWERED, powered), Block.NOTIFY_ALL);
+        }
     }
 
     @Nullable